Address remaining PR review fixes
